+++ src/components/forms/fields/index.ts
@@ -1,6 +1,8 @@
 // Export all form field components
+export * from './FormError';
 export * from './FormInput';
 export * from './FormSelect';
 export * from './FormCheckbox';
 export * from './FormDropdown';
+export * from './FormDateRange';
 export * from './FormBadgesSelection';